# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "camp-backend",
    "description": "\n    # AIエージェントアプリケーション API\n\n    LangChain/LangGraphを使用したAIエージェントアプリケーションのバックエンドAPI。\n\n    ## 認証モード: development\n\n    ### 開発モード認証（モック）\n\n    ## 主な機能\n\n    - **AIエージェントとのチャット**: LangGraphによる高度な会話エンジン\n    - **セッション管理**: 会話履歴の保存と取得\n    - **ファイル処理**: ファイルのアップロード・ダウンロード\n    - **認証・認可**: JWT認証 とロールベース制御\n    - **モニタリング**: Prometheusメトリクス収集\n\n    ## 認証\n\n    ほとんどのエンドポイントはゲストアクセス可能ですが、一部の機能では認証が必要です。\n    認証が必要な場合は、`Authorization: Bearer <token>` ヘッダーを含めてください。\n\n    ## レート制限\n\n    APIリクエストは100リクエスト/分に制限されています。\n    ",
    "contact": {
      "name": "開発チーム",
      "email": "nomura.koshiro@gmail.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.2"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト一覧取得",
        "description": "認証されたユーザーが所属するプロジェクトの一覧を取得します。\n\n    **認証が必要です。**\n\n    - 自分がメンバーとして所属するプロジェクトのみ表示されます\n    - ロール（OWNER/ADMIN/MEMBER/VIEWER）に関係なく全て取得されます\n\n    クエリパラメータ:\n        - skip: スキップするレコード数（デフォルト: 0）\n        - limit: 取得する最大レコード数（デフォルト: 100、最大: 1000）\n        - is_active: アクティブフラグフィルタ（オプション）",
        "operationId": "list_projects_api_v1_projects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "スキップするレコード数",
              "default": 0,
              "title": "Skip"
            },
            "description": "スキップするレコード数"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "取得する最大レコード数",
              "default": 100,
              "title": "Limit"
            },
            "description": "取得する最大レコード数"
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "アクティブフラグフィルタ",
              "title": "Is Active"
            },
            "description": "アクティブフラグフィルタ"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response List Projects Api V1 Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト作成",
        "description": "新しいプロジェクトを作成します。\n\n    **認証が必要です。**\n\n    - 作成者は自動的にOWNERロールとしてプロジェクトメンバーに追加されます\n    - プロジェクトコードは一意である必要があります\n\n    リクエストボディ:\n        - name: プロジェクト名（最大255文字）\n        - code: プロジェクトコード（最大50文字、一意）\n        - description: プロジェクト説明（オプション）",
        "operationId": "create_project_api_v1_projects_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト詳細取得",
        "description": "指定されたIDのプロジェクト情報を取得します。\n\n    **認証が必要です。**\n    **プロジェクトメンバーのみアクセス可能です。**\n\n    パスパラメータ:\n        - project_id: プロジェクトID（UUID形式）",
        "operationId": "get_project_api_v1_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト情報更新",
        "description": "プロジェクト情報を更新します。\n\n    **OWNER/ADMINロールが必要です。**\n\n    更新可能なフィールド:\n        - name: プロジェクト名\n        - description: プロジェクト説明\n        - is_active: アクティブフラグ\n\n    注意:\n        - code（プロジェクトコード）は更新できません\n        - 指定されたフィールドのみが更新されます（PATCH）",
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクト削除",
        "description": "プロジェクトを削除します。\n\n    **OWNERロールが必要です。**\n\n    パスパラメータ:\n        - project_id: プロジェクトID（UUID形式）\n\n    注意:\n        - 削除は物理削除です（データベースから完全に削除されます）\n        - CASCADE設定により、関連する ProjectMember と ProjectFile も自動削除されます\n        - この操作は取り消せません",
        "operationId": "delete_project_api_v1_projects__project_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/code/{code}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "プロジェクトコード検索",
        "description": "指定されたコードのプロジェクト情報を取得します。\n\n    **認証が必要です。**\n    **プロジェクトメンバーのみアクセス可能です。**\n\n    パスパラメータ:\n        - code: プロジェクトコード",
        "operationId": "get_project_by_code_api_v1_projects_code__code__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "プロジェクト名"
          },
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Code",
            "description": "プロジェクトコード（一意識別子）"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "プロジェクト説明"
          }
        },
        "type": "object",
        "required": [
          "name",
          "code"
        ],
        "title": "ProjectCreate",
        "description": "プロジェクト作成リクエストスキーマ。\n\n新規プロジェクト作成時に使用します。\n\nAttributes:\n    name (str): プロジェクト名（ProjectBaseから継承）\n    code (str): プロジェクトコード（ProjectBaseから継承）\n    description (str | None): プロジェクト説明（ProjectBaseから継承）\n\nExample:\n    >>> project = ProjectCreate(\n    ...     name=\"AI Development Project\",\n    ...     code=\"AI-001\",\n    ...     description=\"Project for AI model development\"\n    ... )"
      },
...
      "ProjectResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "プロジェクト名"
          },
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Code",
            "description": "プロジェクトコード（一意識別子）"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "プロジェクト説明"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "プロジェクトID（UUID）"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "アクティブフラグ"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By",
            "description": "作成者のユーザーID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新日時"
          }
        },
        "type": "object",
        "required": [
          "name",
          "code",
          "id",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse",
        "description": "プロジェクト情報レスポンススキーマ。\n\nAPIレスポンスでプロジェクト情報を返す際に使用します。\n\nAttributes:\n    id (uuid.UUID): プロジェクトID（UUID）\n    name (str): プロジェクト名（ProjectBaseから継承）\n    code (str): プロジェクトコード（ProjectBaseから継承）\n    description (str | None): プロジェクト説明（ProjectBaseから継承）\n    is_active (bool): アクティブフラグ\n    created_by (uuid.UUID | None): 作成者のユーザーID\n    created_at (datetime): 作成日時\n    updated_at (datetime): 更新日時\n\nExample:\n    >>> from datetime import UTC\n    >>> project = ProjectResponse(\n    ...     id=uuid.uuid4(),\n    ...     name=\"AI Project\",\n    ...     code=\"AI-001\",\n    ...     description=\"AI development project\",\n    ...     is_active=True,\n    ...     created_by=uuid.uuid4(),\n    ...     created_at=datetime.now(UTC),\n    ...     updated_at=datetime.now(UTC)\n    ... )\n\nNote:\n    - from_attributesを有効にしているため、ORMモデルから直接変換可能です"
      },
...
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "プロジェクト名"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "プロジェクト説明"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "アクティブフラグ"
          }
        },
        "type": "object",
        "title": "ProjectUpdate",
        "description": "プロジェクト更新リクエストスキーマ。\n\nプロジェクト情報の更新時に使用します。\n\nAttributes:\n    name (str | None): プロジェクト名（オプション）\n    description (str | None): プロジェクト説明（オプション）\n    is_active (bool | None): アクティブフラグ（オプション）\n\nExample:\n    >>> update = ProjectUpdate(\n    ...     name=\"Updated Project Name\",\n    ...     description=\"Updated description\"\n    ... )\n\nNote:\n    - すべてのフィールドはオプションです\n    - code は変更できません（一意識別子のため）"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}